+++ travel_api/ViewModels/Requests/EFRequest/FeedbackRequest.cs
@@ -8,7 +8,7 @@
 
         public string FeedbackContent { get; set; }
 
-        public decimal FeedbackRate { get; set; }
+        public int FeedbackRate { get; set; }
 
         public int TripType { get; set; }
